requirements.txt
flask
python-dotenv
git+https://github.com/nedsociety/pysandboxie
pywin32